# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ActivationAccessRead": {
        "additionalProperties": false,
        "properties": {
          "login": {
            "title": "Login",
            "type": "string"
          },
          "recovery_key": {
            "title": "Recovery Key",
            "type": "string"
          }
        },
        "required": [
          "login",
          "recovery_key"
        ],
        "title": "ActivationAccessRead",
        "type": "object"
      },
...
      "BootstrapContact": {
        "additionalProperties": false,
        "properties": {
          "firmware_version": {
            "maxLength": 32,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9._-]+$",
            "title": "Firmware Version",
            "type": "string"
          }
        },
        "required": [
          "firmware_version"
        ],
        "title": "BootstrapContact",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrganizationRead": {
        "description": "Публічне представлення організації в API.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "OrganizationRead",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend API платформи TechBaza IoT Pump Control",
    "title": "TechBaza Backend",
    "version": "0.49.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/bootstrap/{controller_id}/configuration": {
      "post": {
        "operationId": "bootstrap_configuration_api_v1_bootstrap__controller_id__configuration_post",
        "parameters": [
          {
            "in": "path",
            "name": "controller_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Controller Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BootstrapContact"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapConfiguration"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bootstrap Configuration",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/v1/bootstrap/{controller_id}/contact": {
      "post": {
        "operationId": "bootstrap_contact_api_v1_bootstrap__controller_id__contact_post",
        "parameters": [
          {
            "in": "path",
            "name": "controller_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Controller Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BootstrapContact"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bootstrap Contact",
        "tags": [
          "onboarding"
        ]
      }
    },
...
    "/api/v1/connect/organizations": {
      "get": {
        "operationId": "buyer_organizations_api_v1_connect_organizations_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrganizationRead"
                  },
                  "title": "Response Buyer Organizations Api V1 Connect Organizations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Buyer Organizations",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/v1/connect/sites": {
      "get": {
        "operationId": "buyer_sites_api_v1_connect_sites_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SiteRead"
                  },
                  "title": "Response Buyer Sites Api V1 Connect Sites Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Buyer Sites",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/v1/connect/{controller_id}": {
      "get": {
        "operationId": "connection_api_v1_connect__controller_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "controller_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Controller Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Connection",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/v1/connect/{controller_id}/access/{operation}": {
      "post": {
        "operationId": "controller_operation_api_v1_connect__controller_id__access__operation__post",
        "parameters": [
          {
            "in": "path",
            "name": "controller_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Controller Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "operation",
            "required": true,
            "schema": {
              "enum": [
                "rotate",
                "revoke",
                "release"
              ],
              "title": "Operation",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControllerOperation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ControllerStatus"
                    },
                    {
                      "$ref": "#/components/schemas/TransferRead"
                    }
                  ],
                  "title": "Response Controller Operation Api V1 Connect  Controller Id  Access  Operation  Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Controller Operation",
        "tags": [
          "onboarding"
        ]
      }
    },
...
    "/api/v1/connect/{controller_id}/claim": {
      "post": {
        "operationId": "claim_controller_api_v1_connect__controller_id__claim_post",
        "parameters": [
          {
            "in": "path",
            "name": "controller_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Controller Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Claim Controller",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/v1/connect/{controller_id}/equipment": {
      "put": {
        "operationId": "equipment_selection_api_v1_connect__controller_id__equipment_put",
        "parameters": [
          {
            "in": "path",
            "name": "controller_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Controller Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentSelection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Equipment Selection",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/v1/connect/{controller_id}/security": {
      "post": {
        "operationId": "activation_security_api_v1_connect__controller_id__security_post",
        "parameters": [
          {
            "in": "path",
            "name": "controller_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Controller Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivationAccessRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Activation Security",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/v1/connect/{controller_id}/status": {
      "get": {
        "operationId": "lifecycle_status_api_v1_connect__controller_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "controller_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Controller Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControllerStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Lifecycle Status",
        "tags": [
          "onboarding"
        ]
      }
    },
...
    "/api/v1/factory/controllers": {
      "get": {
        "operationId": "factory_list_api_v1_factory_controllers_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FactoryRead"
                  },
                  "title": "Response Factory List Api V1 Factory Controllers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Factory List",
        "tags": [
          "onboarding"
        ]
      },
      "post": {
        "operationId": "factory_create_api_v1_factory_controllers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactorySecrets"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Factory Create",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/v1/factory/controllers/{controller_id}/audit": {
      "get": {
        "operationId": "factory_audit_api_v1_factory_controllers__controller_id__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "controller_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Controller Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Factory Audit",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/v1/factory/controllers/{controller_id}/quarantine": {
      "post": {
        "operationId": "quarantine_controller_api_v1_factory_controllers__controller_id__quarantine_post",
        "parameters": [
          {
            "in": "path",
            "name": "controller_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Controller Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactoryQuarantine"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactoryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Quarantine Controller",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/v1/factory/controllers/{controller_id}/reset": {
      "post": {
        "operationId": "reset_controller_api_v1_factory_controllers__controller_id__reset_post",
        "parameters": [
          {
            "in": "path",
            "name": "controller_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Controller Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactoryReset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactorySecrets"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reset Controller",
        "tags": [
          "onboarding"
        ]
      }
    },
...
  }
}